Moved exceptions and added handling for more of them
